+++ src/include/OpenImageIO/simd.h
@@ -1020,12 +1020,7 @@ public:
     vint4& operator=(int a) { load(a); return *this; }
 
     /// Assignment from another vint4
-#if !defined(__INTEL_COMPILER)
     vint4& operator=(const vint4& other) = default;
-#else
-    // For explanation of the necessity of this, see implementation comment.
-    vint4& operator=(const vint4& other);
-#endif
 
     /// Component access (get)
     int operator[] (int i) const;
@@ -1314,12 +1309,7 @@ public:
     vint8& operator=(int a) { load(a); return *this; }
 
     /// Assignment from another vint8
-#if !defined(__INTEL_COMPILER)
     vint8& operator=(const vint8& other) = default;
-#else
-    // For explanation of the necessity of this, see implementation comment.
-    vint8& operator=(const vint8& other);
-#endif
 
     /// Component access (get)
     int operator[] (int i) const;
@@ -1614,12 +1604,7 @@ public:
     vint16& operator=(int a) { load(a); return *this; }
 
     /// Assignment from another vint16
-#if !defined(__INTEL_COMPILER)
     vint16& operator=(const vint16& other) = default;
-#else
-    // For explanation of the necessity of this, see implementation comment.
-    vint16& operator=(const vint16& other);
-#endif
 
     /// Component access (get)
     int operator[] (int i) const;
@@ -4106,18 +4091,6 @@ OIIO_FORCEINLINE bool none (const vbool16& v) { return reduce_or(v) == false; }
 //////////////////////////////////////////////////////////////////////
 // vint4 implementation
 
-#if defined(__INTEL_COMPILER)
-// For reasons we don't understand, all sorts of failures crop up only on icc
-// if we make this =default. Although we still support icc for now, it's a
-// discontinued compiler, so we special-case it here rather than spend a lot
-// of time investigating what might be broken (and would of course never be
-// fixed if it's a compiler bug).
-OIIO_FORCEINLINE vint4& vint4::operator=(const vint4& other) {
-    m_simd = other.m_simd;
-    return *this;
-}
-#endif
-
 OIIO_FORCEINLINE int vint4::operator[] (int i) const {
     OIIO_DASSERT(i<elements);
     return m_val[i];
@@ -5004,18 +4977,6 @@ OIIO_FORCEINLINE vint4 safe_mod (const vint4& a, int b) {
 //////////////////////////////////////////////////////////////////////
 // vint8 implementation
 
-#if defined(__INTEL_COMPILER)
-// For reasons we don't understand, all sorts of failures crop up only on icc
-// if we make this =default. Although we still support icc for now, it's a
-// discontinued compiler, so we special-case it here rather than spend a lot
-// of time investigating what might be broken (and would of course never be
-// fixed if it's a compiler bug).
-OIIO_FORCEINLINE vint8& vint8::operator=(const vint8& other) {
-    m_simd = other.m_simd;
-    return *this;
-}
-#endif
-
 OIIO_FORCEINLINE int vint8::operator[] (int i) const {
     OIIO_DASSERT(i<elements);
     return m_val[i];
@@ -5833,18 +5794,6 @@ OIIO_FORCEINLINE vint8 safe_mod (const vint8& a, int b) {
 
 //////////////////////////////////////////////////////////////////////
 // vint16 implementation
-
-#if defined(__INTEL_COMPILER)
-// For reasons we don't understand, all sorts of failures crop up only on icc
-// if we make this =default. Although we still support icc for now, it's a
-// discontinued compiler, so we special-case it here rather than spend a lot
-// of time investigating what might be broken (and would of course never be
-// fixed if it's a compiler bug).
-OIIO_FORCEINLINE vint16& vint16::operator=(const vint16& other) {
-    m_simd = other.m_simd;
-    return *this;
-}
-#endif
 
 OIIO_FORCEINLINE int vint16::operator[] (int i) const {
     OIIO_DASSERT(i<elements);
@@ -10306,15 +10255,6 @@ template<> struct fmt::formatter<OIIO::simd::vint16>
 template<> struct fmt::formatter<OIIO::simd::matrix44>
     : OIIO::pvt::array_formatter<OIIO::simd::matrix44, float, 16> {};
 
-
-// Allow C++ metaprogramming to understand that the simd types are trivially
-// copyable (i.e. memcpy to copy simd types is fine).
-#if defined(__INTEL_COMPILER)
-// Necessary because we have to define the vint types copy constructors on icc
-template<> struct std::is_trivially_copyable<OIIO::simd::vint4> : std::true_type {};
-template<> struct std::is_trivially_copyable<OIIO::simd::vint8> : std::true_type {};
-template<> struct std::is_trivially_copyable<OIIO::simd::vint16> : std::true_type {};
-#endif
 
 
 #undef SIMD_DO