+++ src/libOpenImageIO/formatspec.cpp
@@ -41,7 +41,6 @@ inline void
 get_default_quantize_(long long& quant_min, long long& quant_max) noexcept
 {
     OIIO_PRAGMA_WARNING_PUSH
-    OIIO_INTEL_PRAGMA(warning disable 173)
     if (std::numeric_limits<T>::is_integer) {
         quant_min = (long long)std::numeric_limits<T>::min();
         quant_max = (long long)std::numeric_limits<T>::max();